+++ lib/cmd.js
@@ -1,5 +1,6 @@
 const help = require('./cmd/help')
 const init = require('./cmd/init')
+const reset = require('./cmd/reset')
 const test = require('./cmd/test')
 const destroy = require('./cmd/destroy')
 const newvers = require('./cmd/new')
@@ -12,6 +13,7 @@ const launch = function (cmd, params) {
     switch (cmd) {
         case 'help': help(params); break
         case 'init': init(params); break
+        case 'reset': reset(params); break
         case 'test': test(params); break
         case '--system-get': system.get(params); break
         case '--system-set': system.set(params); break
