+++ lib/cmd/help.js
@@ -7,9 +7,6 @@ const figlet = require('figlet')
 module.exports = function () {
     clear();
 
-    console.log(
-        chalk.green(
-            figlet.textSync('VERS', {horizontalLayout: 'full'})
-        )
-    )
+    console.log(chalk.green(figlet.textSync('VERS', {horizontalLayout: 'full'})))
+    console.log('Wellcome to vers help ! Thanks for using the system vers !')
 }
