+++ lib/system.js
@@ -34,6 +34,8 @@ function configSystem (config, rm = false) {
         last = data.configs.find(c => c.name === config.name || c.path === config.path)
         if (!last) data.configs.push(config)
         last = config
+        const index = data.configs.findIndex(c => c.path === last.path)
+        if (index >= 0) data.configs[index] = last
     }
     check(data)
     return last
